+++ Airship/AirshipCore/Tests/AirshipLocaleManagerTest.swift
@@ -7,18 +7,24 @@ import AirshipCore
 
 final class AirshipLocaleManagerTest: XCTestCase {
 
-    private let notificaitonCenter: AirshipNotificationCenter = AirshipNotificationCenter(
+    private let notificationCenter: AirshipNotificationCenter = AirshipNotificationCenter(
         notificationCenter: NotificationCenter()
     )
     private let dataStore: PreferenceDataStore = PreferenceDataStore(
         appKey: UUID().uuidString
     )
+    private var config: RuntimeConfig!
     private var localeManager: AirshipLocaleManager!
 
     override func setUpWithError() throws {
+        config = RuntimeConfig(
+                   config: AirshipConfig(),
+                   dataStore: dataStore
+               )
         localeManager = AirshipLocaleManager(
             dataStore: dataStore,
-            notificationCenter: notificaitonCenter
+            config: config,
+            notificationCenter: notificationCenter
         )
     }
 
@@ -36,10 +42,23 @@ final class AirshipLocaleManagerTest: XCTestCase {
         localeManager.clearLocale()
         XCTAssertEqual(self.localeManager.currentLocale, Locale.autoupdatingCurrent)
     }
+    
+    func testLocaleWithUseUserPreferredLocale() throws {
+        config.useUserPreferredLocale = true
+        let preferredLocale = Locale(identifier: Locale.preferredLanguages[0])
+        XCTAssertEqual(self.localeManager.currentLocale, preferredLocale)
+        
+        let french = Locale(identifier: "fr")
+        localeManager.currentLocale = french
+        XCTAssertEqual(self.localeManager.currentLocale, french)
+        
+        localeManager.clearLocale()
+        XCTAssertEqual(self.localeManager.currentLocale, preferredLocale)
+    }
 
     func testNotificationWhenOverrideChanges() {
         let expectation = self.expectation(description: "update called")
-        self.notificaitonCenter.addObserver(
+        self.notificationCenter.addObserver(
             forName: AirshipLocaleManager.localeUpdatedEvent
         ) { _ in
             expectation.fulfill()
@@ -54,7 +73,7 @@ final class AirshipLocaleManagerTest: XCTestCase {
         localeManager.currentLocale = Locale(identifier: "fr")
 
         let expectation = self.expectation(description: "update called")
-        self.notificaitonCenter.addObserver(
+        self.notificationCenter.addObserver(
             forName: AirshipLocaleManager.localeUpdatedEvent
         ) { _ in
             expectation.fulfill()
@@ -67,13 +86,13 @@ final class AirshipLocaleManagerTest: XCTestCase {
 
     func testNotificationWhenAutoUpdateChanges() {
         let expectation = self.expectation(description: "update called")
-        self.notificaitonCenter.addObserver(
+        self.notificationCenter.addObserver(
             forName: AirshipLocaleManager.localeUpdatedEvent
         ) { _ in
             expectation.fulfill()
         }
 
-        self.notificaitonCenter.post(name: NSLocale.currentLocaleDidChangeNotification)
+        self.notificationCenter.post(name: NSLocale.currentLocaleDidChangeNotification)
 
         self.waitForExpectations(timeout: 10.0)
     }